+++ eras/dijkstra/testlib/Test/Cardano/Ledger/Dijkstra/TreeDiff.hs
@@ -1,7 +1,9 @@
 {-# LANGUAGE FlexibleInstances #-}
 {-# OPTIONS_GHC -Wno-orphans #-}
 
-module Test.Cardano.Ledger.Dijkstra.TreeDiff () where
+module Test.Cardano.Ledger.Dijkstra.TreeDiff (
+  module Test.Cardano.Ledger.Conway.TreeDiff,
+) where
 
 import Cardano.Ledger.Dijkstra (DijkstraEra)
 import Cardano.Ledger.Dijkstra.Core (EraTxBody (..), PlutusScript)