+++ eras/mary/impl/test/Test/Cardano/Ledger/Mary/BinarySpec.hs
@@ -6,9 +6,7 @@
 module Test.Cardano.Ledger.Mary.BinarySpec (spec) where
 
 import Cardano.Ledger.Mary
-import Data.Default (def)
 import Test.Cardano.Ledger.Common
-import Test.Cardano.Ledger.Core.Binary (specUpgrade)
 import Test.Cardano.Ledger.Core.Binary as Binary (decoderEquivalenceCoreEraTypesSpec, txSizeSpec)
 import Test.Cardano.Ledger.Core.Binary.RoundTrip (RuleListEra (..))
 import Test.Cardano.Ledger.Mary.Arbitrary ()
@@ -18,7 +16,6 @@ import Test.Cardano.Ledger.Shelley.Binary.RoundTrip (roundTripShelleyCommonSpec)
 
 spec :: Spec
 spec = do
-  specUpgrade @MaryEra def
   describe "RoundTrip" $ do
     roundTripShelleyCommonSpec @MaryEra
   describe "DecCBOR instances equivalence" $ do