+++ libs/cardano-ledger-api/src/Cardano/Ledger/Api/Tx/AuxData.hs
@@ -2,7 +2,6 @@ module Cardano.Ledger.Api.Tx.AuxData (
   EraTxAuxData (TxAuxData),
   mkBasicTxAuxData,
   metadataTxAuxDataL,
-  upgradeTxAuxData,
   hashTxAuxData,
   validateTxAuxData,
 
@@ -21,6 +20,10 @@ module Cardano.Ledger.Api.Tx.AuxData (
   AlonzoTxAuxData (..),
   mkAlonzoTxAuxData,
   getAlonzoTxAuxDataScripts,
+
+  -- * Upgrade
+  binaryUpgradeTxAuxData,
+  upgradeTxAuxData,
 ) where
 
 import Cardano.Ledger.Allegra.TxAuxData (AllegraEraTxAuxData (..), AllegraTxAuxData (..))
@@ -30,6 +33,6 @@ import Cardano.Ledger.Alonzo.TxAuxData (
   getAlonzoTxAuxDataScripts,
   mkAlonzoTxAuxData,
  )
-import Cardano.Ledger.Api.Era ()
-import Cardano.Ledger.Core (EraTxAuxData (..), hashTxAuxData)
+import Cardano.Ledger.Api.Era (EraApi (..))
+import Cardano.Ledger.Core (EraTxAuxData (..), binaryUpgradeTxAuxData, hashTxAuxData)
 import Cardano.Ledger.Shelley.TxAuxData (Metadatum (..), ShelleyTxAuxData (..))